Moved server files for submission component
