+++ packages/component-submission/server/helpers/manuscriptInputValidationSchema.js
