+++ packages/component-submission/server/resolvers/removeSupportingFiles.js
