+++ packages/component-submission/server/resolvers/removeSupportingFiles.test.js
