+++ packages/component-submission/server/resolvers/removeUploadedManuscript.js
