+++ packages/component-submission/server/resolvers/submitManuscript.js
