+++ packages/component-submission/server/resolvers/updateManuscript.js
