+++ packages/component-submission/server/resolvers/updateManuscript.test.js
@@ -10,7 +10,7 @@ const {
   manuscriptInput,
 } = require('./index.test.data')
 
-const replaySetup = require('../../../../../test/helpers/replay-setup')
+const replaySetup = require('../../../../test/helpers/replay-setup')
 
 describe('Manuscript resolvers', () => {
   const profileId = userData.identities[0].identifier
