+++ packages/component-submission/server/resolvers/uploadManuscript.js
