+++ packages/component-submission/server/resolvers/uploadSupportingFile.js
